08_download_cmems/src/requirements.txt
motuclient==1.8.8
python-dotenv==0.15.0
